+++ src/encode.rs
@@ -2,7 +2,9 @@ use crate::Constants;
 use crate::FrameType;
 #[cfg(feature = "std")]
 use crate::Slipmux;
-use crate::checksum::{CONF_FCS16, fcs16_finish, fcs16_part};
+use crate::checksum::CONF_FCS16;
+use crate::checksum::fcs16_finish;
+use crate::checksum::fcs16_part;
 use serial_line_ip::EncodeTotals;
 use serial_line_ip::Encoder;
 
@@ -32,18 +34,30 @@ pub fn encode_packet(packet: &[u8], buffer: &mut [u8]) -> usize {
 #[cfg(feature = "std")]
 #[must_use]
 pub fn encode_buffered(input: Slipmux) -> Vec<u8> {
+    // Calculate the worst-case amount of space needed to encode this much data
+    const fn space_requirement(data_len: usize) -> usize {
+        const FRAMETYPE_BYTE: usize = 1;
+        const ENDFRAME_BYTE: usize = 1;
+        // Assuming every single byte needs to be escaped
+        let max_encoded_len: usize = data_len * 2;
+        // leading endframe byte to flush privious unclean states (if any)
+        // see slip rfc
+        ENDFRAME_BYTE + FRAMETYPE_BYTE + max_encoded_len + ENDFRAME_BYTE
+    }
     let mut buffer: Vec<u8> = vec![];
     let length = match input {
         Slipmux::Diagnostic(s) => {
-            buffer.resize(s.len() * 2, 0);
+            buffer.resize(space_requirement(s.len()), 0);
             encode(FrameType::Diagnostic, s.as_bytes(), &mut buffer)
         }
         Slipmux::Configuration(conf) => {
-            buffer.resize(conf.len() * 2 + 2, 0);
+            const CHECKSUM_BYTES: usize = 2;
+            buffer.resize(space_requirement(conf.len() + CHECKSUM_BYTES), 0);
             encode(FrameType::Configuration, &conf, &mut buffer)
         }
         Slipmux::Packet(packet) => {
-            buffer.resize(packet.len() * 2, 0);
+            // IP packets don't have a start byte so this is over provisioning one byte
+            buffer.resize(space_requirement(packet.len()), 0);
             encode(FrameType::Ip, &packet, &mut buffer)
         }
     };
@@ -183,5 +197,67 @@ mod tests {
                 Constants::END
             ]
         );
+    }
+
+    #[test]
+    #[cfg(feature = "std")]
+    fn encode_buffered_max_encoding_size() {
+        let mut data: Vec<u8> = vec![
+            Constants::END,
+            Constants::ESC,
+            Constants::ESC,
+            Constants::END,
+            Constants::ESC,
+            Constants::END,
+            Constants::ESC,
+            Constants::END,
+        ];
+        for _ in 0..=45 {
+            data.push(Constants::ESC);
+        }
+        assert_eq!(data.len(), 54);
+        let input = Slipmux::Configuration(data);
+        let result = encode_buffered(input);
+
+        // endbyte + startbyte + data * 2 + fcs * 2 + endbyte
+        // data * 2 because every single byte needs escaping
+        // fcs * 2 because both fcs bytes need escaping
+        assert_eq!(result.len(), 1 + 1 + 54 * 2 + 2 * 2 + 1);
+
+        let mut expected: Vec<u8> = vec![
+            Constants::END,
+            Constants::CONFIGURATION,
+            Constants::ESC,
+            Constants::ESC_END,
+            Constants::ESC,
+            Constants::ESC_ESC,
+            Constants::ESC,
+            Constants::ESC_ESC,
+            Constants::ESC,
+            Constants::ESC_END,
+            Constants::ESC,
+            Constants::ESC_ESC,
+            Constants::ESC,
+            Constants::ESC_END,
+            Constants::ESC,
+            Constants::ESC_ESC,
+            Constants::ESC,
+            Constants::ESC_END,
+        ];
+        for _ in 0..=45 {
+            expected.push(Constants::ESC);
+            expected.push(Constants::ESC_ESC);
+        }
+
+        // Checksum is 0xC0C0
+        expected.push(Constants::ESC);
+        expected.push(Constants::ESC_END);
+        expected.push(Constants::ESC);
+        expected.push(Constants::ESC_END);
+
+        // End frame
+        expected.push(Constants::END);
+
+        assert_eq!(result, expected);
     }
 }